refactor: 리뷰사항 반영

+++ apps/dragons_api/src/main/java/com/dragons/interfaces/api/payment/dto/PaymentV1Dto.java
@@ -33,7 +33,7 @@ public class PaymentV1Dto {
         @Schema(description = "소유자명", example = "홍길동") @NotBlank String cardholderName,
 
         // amount must be > 0
-        @Schema(description = "결제 금액(양수)", example = "10000") @Positive(message = "금액은 양수여야합니다.") int amount,
+        @Schema(description = "결제 금액(양수)", example = "10000") @Positive(message = "금액은 양수여야합니다.") long amount,
 
         @Schema(description = "요금제/플랜 타입", example = "premium")
         @Pattern(regexp = "^(premium|basic)$", message = "유효한 플랜 타입이 아닙니다.")
@@ -50,7 +50,7 @@ public class PaymentV1Dto {
     @Schema(name = "PaymentCardResponse", description = "카드 결제 응답")
     public record Response(
         @Schema(description = "마스킹된 카드 번호", example = "123456******5678") String cardNumber, // 카드 번호
-        @Schema(description = "결제 금액", example = "100") int amount, // 금액
+        @Schema(description = "결제 금액", example = "100") long amount, // 금액
         @Schema(description = "요금제/플랜 타입", example = "premium") String planType // 결제타입
     ) {
 
@@ -67,7 +67,7 @@ public class PaymentV1Dto {
         @Schema(description = "은행 코드(3자리)", example = "004") @NotBlank @Pattern(regexp = "\\d{3}", message = "은행 코드는 3자리 숫자입니다.") String bankCode,
         @Schema(description = "계좌번호(10~14자리)", example = "123456789012") @NotBlank @Pattern(regexp = "\\d{10,14}", message = "계좌번호는 10~14자리 숫자입니다.") String accountNumber,
         @Schema(description = "예금주명(2~20자)", example = "홍길동") @NotBlank @Size(min = 2, max = 20, message = "예금주명은 2~20자리입니다.") String depositorName,
-        @Schema(description = "결제 금액(양수)", example = "10000") @Positive(message = "금액은 양수여야합니다.") int amount,
+        @Schema(description = "결제 금액(양수)", example = "10000") @Positive(message = "금액은 양수여야합니다.") long amount,
         @Schema(description = "요금제/플랜 타입", example = "premium")
         @Pattern(regexp = "^(premium|basic)$", message = "유효한 플랜 타입이 아닙니다.")
         String planType) {
